Finally added Persistent Cave States to the GDO Framework You may now save any set of sections / app deployments and restore it to/from json object stored under /states/ on the server.
Next step - store the state of individual apps aswell - e.g. which image is displayed.

+++ GDO/Utility/Utilities.cs
@@ -132,7 +132,9 @@ namespace GDO.Utility
             try
             {
                 string path = Directory.GetCurrentDirectory() + @"\\" + folderName + "\\" + fileName + ".json";
-                File.Delete(path);
+                if (File.Exists(path)) {
+                    File.Delete(path);
+                }
             }
             catch (Exception e)
             {
